requirements.txt
paho-mqtt==1.5.0
pvporcupine~=1.9.0
rhasspy-hermes~=0.6.0
